changes schemat for trip logs miles from an integer to a float

+++ db/migrate/20190218182704_create_trip_log.rb
@@ -2,7 +2,7 @@ class CreateTripLog < ActiveRecord::Migration[5.2]
   def change
     create_table :trip_logs do |t|
       t.string :title
-      t.integer :miles
+      t.float :miles
       t.string :description
       t.references :user, foreign_key: true
     end
